rename template files
