+++ frontend/packages/kubevirt-plugin/src/models/templates/vm-template-yaml.ts
