+++ frontend/packages/kubevirt-plugin/src/models/templates/vm-yaml.ts
